+++ gar/include/aligator/gar/utils.hpp
@@ -3,12 +3,53 @@
 
 #include "lqr-problem.hpp"
 #include "aligator/third-party/boost/core/span.hpp"
+#include <Eigen/SparseCore>
 
-namespace aligator {
-namespace gar {
+namespace aligator::gar {
+namespace helpers {
+/// @brief Helper to assign a dense matrix into a range of coefficients of a
+/// sparse matrix.
+template <typename InType, typename OutScalar>
+void sparseAssignDenseBlock(Eigen::Index i0, Eigen::Index j0,
+                            const Eigen::DenseBase<InType> &input,
+                            Eigen::SparseMatrix<OutScalar> &out, bool update) {
+  assert(i0 + input.rows() <= out.rows() && "Inconsistent rows");
+  assert(j0 + input.cols() <= out.cols() && "Inconsistent cols");
+  using Eigen::Index;
+  for (Index i = 0; i < input.rows(); i++) {
+    for (Index j = 0; j < input.cols(); j++) {
+      if (update)
+        out.coeffRef(i0 + i, j0 + j) = input(i, j);
+      else
+        out.insert(i0 + i, j0 + j) = input(i, j);
+    }
+  }
+}
 
 template <typename Scalar>
-auto lqrComputeKktError(
+void sparseAssignDiagonal(Eigen::Index i0, Eigen::Index i1, Scalar value,
+                          Eigen::SparseMatrix<Scalar> &out, bool update) {
+  using Eigen::Index;
+  assert(i0 <= i1 && "i0 should be lesser than i1. Can't assign empty range.");
+  assert(i1 <= out.rows() && "Inconsistent rows");
+  assert(i1 <= out.cols() && "Inconsistent cols");
+  for (Index kk = i0; kk < i1; kk++) {
+    if (update)
+      out.coeffRef(kk, kk) = value;
+    else
+      out.insert(kk, kk) = value;
+  }
+}
+} // namespace helpers
+
+template <typename Scalar>
+void lqrCreateSparseMatrix(const LQRProblemTpl<Scalar> &problem,
+                           const Scalar mudyn, const Scalar mueq,
+                           Eigen::SparseMatrix<Scalar> &mat,
+                           Eigen::Matrix<Scalar, -1, 1> &rhs, bool update);
+
+template <typename Scalar>
+std::array<Scalar, 3> lqrComputeKktError(
     const LQRProblemTpl<Scalar> &problem,
     boost::span<const typename math_types<Scalar>::VectorXs> xs,
     boost::span<const typename math_types<Scalar>::VectorXs> us,
@@ -16,7 +57,18 @@ auto lqrComputeKktError(
     boost::span<const typename math_types<Scalar>::VectorXs> lbdas,
     const Scalar mudyn, const Scalar mueq,
     const std::optional<typename math_types<Scalar>::ConstVectorRef> &theta_,
-    bool verbose = false) {
+    bool verbose = false);
+
+template <typename Scalar>
+std::array<Scalar, 3> lqrComputeKktError(
+    const LQRProblemTpl<Scalar> &problem,
+    boost::span<const typename math_types<Scalar>::VectorXs> xs,
+    boost::span<const typename math_types<Scalar>::VectorXs> us,
+    boost::span<const typename math_types<Scalar>::VectorXs> vs,
+    boost::span<const typename math_types<Scalar>::VectorXs> lbdas,
+    const Scalar mudyn, const Scalar mueq,
+    const std::optional<typename math_types<Scalar>::ConstVectorRef> &theta_,
+    bool verbose) {
   fmt::print("[{}] ", __func__);
   uint N = (uint)problem.horizon();
   using VectorXs = typename math_types<Scalar>::VectorXs;
@@ -111,78 +163,7 @@ auto lqrComputeKktError(
 template <typename Scalar>
 bool lqrDenseMatrix(const LQRProblemTpl<Scalar> &problem, Scalar mudyn,
                     Scalar mueq, typename math_types<Scalar>::MatrixXs &mat,
-                    typename math_types<Scalar>::VectorXs &rhs) {
-  using knot_t = LQRKnotTpl<Scalar>;
-  const auto &knots = problem.stages;
-  const size_t N = size_t(problem.horizon());
-
-  if (!problem.isInitialized())
-    return false;
-
-  const uint nrows = lqrNumRows(problem);
-  mat.conservativeResize(nrows, nrows);
-  rhs.conservativeResize(nrows);
-  mat.setZero();
-
-  uint idx = 0;
-  {
-    const uint nc0 = problem.nc0();
-    const uint nx0 = knots[0].nx;
-    mat.block(nc0, 0, nx0, nc0) = problem.G0.transpose();
-    mat.block(0, nc0, nc0, nx0) = problem.G0;
-    mat.topLeftCorner(nc0, nc0).diagonal().setConstant(-mudyn);
-
-    rhs.head(nc0) = problem.g0;
-    idx += nc0;
-  }
-
-  for (size_t t = 0; t <= N; t++) {
-    const knot_t &model = knots[t];
-    // get block for current variables
-    const uint n = model.nx + model.nu + model.nc;
-    auto block = mat.block(idx, idx, n, n);
-    auto rhsblk = rhs.segment(idx, n);
-    auto Q = block.topLeftCorner(model.nx, model.nx);
-    auto St = block.leftCols(model.nx).middleRows(model.nx, model.nu);
-    auto R = block.block(model.nx, model.nx, model.nu, model.nu);
-    auto C = block.bottomRows(model.nc).leftCols(model.nx);
-    auto D = block.bottomRows(model.nc).middleCols(model.nx, model.nu);
-    auto dual = block.bottomRightCorner(model.nc, model.nc).diagonal();
-    dual.array() = -mueq;
-
-    Q = model.Q;
-    St = model.S.transpose();
-    R = model.R;
-    C = model.C;
-    D = model.D;
-
-    block = block.template selfadjointView<Eigen::Lower>();
-
-    rhsblk.head(model.nx) = model.q;
-    rhsblk.segment(model.nx, model.nu) = model.r;
-    rhsblk.tail(model.nc) = model.d;
-
-    // fill in dynamics
-    // row contains [A; B; 0; -mu*I, E] -> nx + nu + nc + nx + nx2 cols
-    if (t != N) {
-      uint ncols = model.nx + model.nx2 + n;
-      auto row = mat.block(idx + n, idx, model.nx, ncols);
-      row.leftCols(model.nx) = model.A;
-      row.middleCols(model.nx, model.nu) = model.B;
-      row.middleCols(n, model.nx).diagonal().array() = -mudyn;
-      row.rightCols(model.nx) = model.E;
-
-      rhs.segment(idx + n, model.nx2) = model.f;
-
-      auto col = mat.transpose().block(idx + n, idx, model.nx, ncols);
-      col = row;
-
-      // shift by size of block + costate size (nx2)
-      idx += n + model.nx2;
-    }
-  }
-  return true;
-}
+                    typename math_types<Scalar>::VectorXs &rhs);
 
 /// @brief Compute the number of rows in the problem matrix.
 template <typename Scalar>
@@ -286,7 +267,12 @@ auto lqrInitializeSolution(const LQRProblemTpl<Scalar> &problem) {
 }
 
 #ifdef ALIGATOR_ENABLE_TEMPLATE_INSTANTIATION
-extern template auto lqrComputeKktError<context::Scalar>(
+extern template void lqrCreateSparseMatrix<context::Scalar>(
+    const LQRProblemTpl<context::Scalar> &problem, const context::Scalar mudyn,
+    const context::Scalar mueq, Eigen::SparseMatrix<context::Scalar> &mat,
+    context::VectorXs &rhs, bool update);
+extern template std::array<context::Scalar, 3>
+lqrComputeKktError<context::Scalar>(
     const LQRProblemTpl<context::Scalar> &,
     boost::span<const context::VectorXs>, boost::span<const context::VectorXs>,
     boost::span<const context::VectorXs>, boost::span<const context::VectorXs>,
@@ -297,5 +283,4 @@ lqrDenseMatrix<context::Scalar>(const LQRProblemTpl<context::Scalar> &,
                                 context::Scalar, context::Scalar);
 #endif
 
-} // namespace gar
-} // namespace aligator
+} // namespace aligator::gar